Extract confirmation code to a hook
